requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic[email]
PyJWT